made DeletedTripNotifier component
